% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_parameters}
\alias{add_parameters}
\title{Add parameters to an experiment}
\usage{
add_parameters(experiment, pars)
}
\arguments{
\item{experiment}{Name of the experiment}

\item{pars}{data.frame of the parameters.  Column names must match
the existing parameters}
}
\description{
Add parameters to an existing experiment
}

