% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_task}
\alias{add_task}
\title{Add a task to an experiment}
\usage{
add_task(experiment, task, function_name, common_parameters = NULL,
  packages = NULL, scripts = NULL, overwrite = FALSE, purge = FALSE,
  depends = NULL)
}
\arguments{
\item{experiment}{name}

\item{task}{task name}

\item{function_name}{name of function to call}

\item{common_parameters}{Parameters that are constant for all runs within this task}

\item{packages}{Optional character vector of packages to load for
every task in this experiment}

\item{scripts}{Optional characte vector of scripts to source for
every task in this experiment}

\item{overwrite}{Replace existing specifiction, if it exists}

\item{purge}{}

\item{depends}{Does this task depend on other tasks existing}
}
\description{
Adds a task to an existing experiment
}

