% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batteries.R
\docType{data}
\name{batteries}
\alias{batteries}
\title{Data from Rutherford's experiment with batteries}
\format{
A data frame with 30 rows and 5 variables:
\describe{
  \item{treatment}{Treatment indicator. 1 = treatment, 0 = control}
  \item{one}{Outcome variable denoting first bounce time}
  \item{two}{Outcome variable denoting second bounce time}
  \item{three}{Outcome variable denoting third bounce time}
  \item{pre}{Pre-treatment covariate denoting pre-test bounce time}
}
}
\usage{
batteries
}
\description{
This dataset is used in the exercises of Chapter 3 from the textbook.
}
\keyword{datasets}
