% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruneau.R
\docType{data}
\name{bruneau}
\alias{bruneau}
\title{Data from "The power of being heard: The benefits of ‘perspective-giving’ in the context of intergroup conflict"}
\format{
A data frame with 123 rows and 7 variables:
\describe{
  \item{FTar}{Attitude towards Arabs at time 0}
  \item{FTar_1}{Attitude towards Arabs at time 1}
  \item{FTar_2}{Attitude towards Arabs at time 2}
  \item{Average}{Average outcome at time 0}
  \item{Average_1}{Average outcome at time 1}
  \item{group}{Ethnic group. 1 = White, 2 = Mexican}
  \item{condition}{Condition indicator. 2 = receiver, 1 = sender}
}
}
\source{
Bruneau, Emile G., and Rebecca Saxe. “The Power of Being Heard: The Benefits of ‘Perspective-Giving’ in the Context of Intergroup Conflict.” Journal of Experimental Social Psychology, vol. 48, no. 4, 2012, pp. 855–866., https://doi.org/10.1016/j.jesp.2012.02.017.
}
\usage{
bruneau
}
\description{
This dataset is used in the exercises of Chapter 4 from the textbook.
}
\keyword{datasets}
