% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildOutputPath.R
\name{buildOutputPath}
\alias{buildOutputPath}
\title{Wrapper around sprintf and dir.create.}
\usage{
buildOutputPath(root, format.string, ...)
}
\arguments{
\item{root}{[\code{character(1)}]\cr
Root path.}

\item{format.string}{[\code{character(1)}]\cr
Format string for \code{sprintf}.}

\item{...}{[any]\cr
Further atomic values passed down as arguments to \code{sprintf}.}
}
\value{
[\code{invisible(character(1))}] Silently returns the build path.
As a side effect the path is generated.
}
\description{
Wrapper around sprintf and dir.create.
}
