% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copy_runner}
\alias{copy_runner}
\title{Copy helper script}
\usage{
copy_runner(dest = "run_experiment.R")
}
\arguments{
\item{dest}{Filename of the script}
}
\description{
Copy a helper script that runs a parameter set from the experiment.
}

