% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulation_r.R
\name{do_simulation_r}
\alias{do_simulation_r}
\title{Do a simulation}
\usage{
do_simulation_r(csv_filename)
}
\arguments{
\item{csv_filename}{A CSV filename}
}
\value{
Nothing
}
\description{
Do a simulation
}

