% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropExtension}
\alias{dropExtension}
\title{Drop files extension(s).}
\usage{
dropExtension(fn, levels = 1L)
}
\arguments{
\item{fn}{[\code{character}]
Path.}

\item{levels}{[\code{integer(1)}]\cr
How many extensions should be removed?
Defaults to 1.}
}
\description{
Drop files extension(s).
}
