% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frijters.R
\docType{data}
\name{frijters}
\alias{frijters}
\title{Data from "Roommate effects in health outcomes"}
\format{
A data frame with 214 rows and 3 variables:
\describe{
  \item{Roommate_BMI}{Roommate's BMI}
  \item{Past_BMI}{Participant's BMI at baseline}
  \item{Current_BMI}{Participant's BMI after living with roommate for one year}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.1002/hec.3901}
}
\usage{
frijters
}
\description{
This dataset is used in the exercises of Chapter 7 from the textbook.
}
\keyword{datasets}
