% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ids_incomplete}
\alias{ids_incomplete}
\title{Identify incomplete ids}
\usage{
ids_incomplete(experiment, task, test = NULL)
}
\arguments{
\item{experiment}{Name of the experiment}

\item{task}{Name of the task}

\item{test}{Function to apply to each output to determine
"done-ness" (not yet implemented)}
}
\description{
Look for jobs that have not completed.  Very simple minded and
only looks for existance of the output file.  Eventually this will
support running a custom function over the output.
}

