% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProject.R
\name{initProject}
\alias{initProject}
\title{Build up my project folder structure.}
\usage{
initProject(dir)
}
\arguments{
\item{dir}{[\code{character(1)}]\cr
Root directory of the project.}
}
\value{
Nothing
}
\description{
Build up my project folder structure.
}
