% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectResultFolder.R
\name{inspectResultFolder}
\alias{inspectResultFolder}
\title{Glimpse into results folder.}
\usage{
inspectResultFolder(path, pattern = NULL, ...)
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path to result folder.}

\item{pattern}{[\code{character(1)}]\cr
Pattern used to filter files recursively. See argument \code{pattern} of
\code{\link[base]{list.files}}.}

\item{...}{[any]\cr
Passed down to \code{\link{list.files}}.}
}
\value{
Nothing
}
\description{
Glimpse into results folder.
}
