% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_pbs}
\alias{launch_pbs}
\title{Launch jobs via qsub}
\usage{
launch_pbs(experiment, task, id = NULL, jobfile = "pbs_jobs.csv")
}
\arguments{
\item{experiment}{Name of the experiment}

\item{task}{Name of the task}

\item{id}{Vector of ids}

\item{jobfile}{Name of the file to store information associating
pbs job id with experimentrs information.  This is used by
\code{\link{move_pbs_logs}}.}
}
\description{
Launch jobs via qsub, saving information about PBS job numbers
}

