% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_output}
\alias{load_output}
\title{Load output for a given task}
\usage{
load_output(experiment, task, id = NULL, f = identity, missing_ok = TRUE)
}
\arguments{
\item{experiment}{Name of the experiment}

\item{task}{Name of the task}

\item{id}{Optional vector of ids}

\item{f}{Optional Function to apply to each output}

\item{missing_ok}{Logical indicating if missing ids are OK: if so
a \code{NULL} will be given for this set.}
}
\description{
Load output for a given task
}
\author{
Rich FitzJohn
}

