% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log.default}
\alias{log.default}
\title{Log arbitrary metrics provided as key-value pairs}
\usage{
\method{log}{default}(metric_name, metric, experiment)
}
\arguments{
\item{metric_name}{A character vector specifying the names of the metrics you
want to log. If you provide a character string this name will be used for
all metric results.}

\item{metric}{A numeric vector of the same length as metric_name giving the
result corresponding to the metric defined in metric_name}

\item{experiment}{An experiment object to store your metrics in}
}
\description{
Log arbitrary metrics provided as key-value pairs
}
