% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log}
\alias{log}
\title{Log model run output}
\usage{
log(x, ...)
}
\description{
This function can be used to track model runs from caret train objects. You can
also pass in a metric and values as a key value pair to log arbitrary model
results.
}
