% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log.train}
\alias{log.train}
\title{Log output from caret train objects}
\usage{
\method{log}{train}(caret_train, experiment, description = NULL)
}
\arguments{
\item{caret_train}{A train object from caret}

\item{experiment}{An experiment object}

\item{description}{An optional character string describing the experiment run}
}
\description{
This function is a convenience logging wrapper for caret train objects.
It will automatically extract the following information:
  - model method
  - model type
  - model label
  - control settings used (e.g. 10x3 repeated-cv)
  - tuning results
  - resampling results for final model
}
