% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynn.R
\docType{data}
\name{lynn}
\alias{lynn}
\title{Data from "Effect of Server Posture on Restaurant Tipping"}
\format{
A data frame with 148 rows and 7 variables:
\describe{
  \item{groupsize}{Number of customers dining at table}
  \item{bill}{Total bill amount in dollars}
  \item{tip}{Tip amount in dollars}
  \item{daytime}{Day or evening shift}
  \item{female}{Whether the bill was paid by male or female. 1 = female, 0 = male}
  \item{crouch}{1 = crouch, 0 = standing}
  \item{paid_by_credit_card}{Whether the bill was paid by credit card. 1 = credit card, 0 = not credit card}
}
}
\source{
Lynn, Michael, and Kirby Mynier. “Effect of Server Posture on Restaurant Tipping.” Journal of Applied Social Psychology, vol. 23, no. 8, 1993, pp. 678–685., https://doi.org/10.1111/j.1559-1816.1993.tb01109.x.
}
\usage{
lynn
}
\description{
This dataset is used in the exercises of Chapter 7 from the textbook.
}
\keyword{datasets}
