% Generated by roxygen2 (4.0.2): do not edit by hand
\name{main}
\alias{main}
\title{Run an experiment, from the command line}
\usage{
main(args = NULL)
}
\arguments{
\item{args}{Arguments (or NULL to read from R's \code{commandArgs})}
}
\description{
Run an experiment, from the command line.
}
\details{
Don't use this in interactive mode this won't work
}

