% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_pbs_file}
\alias{make_pbs_file}
\title{Generate pbs files from a template}
\usage{
make_pbs_file(experiment, task, id = NULL,
  email = getOption("experimentr.email"), walltime = "48:00",
  queue = "normal", template = NULL, path = ".")
}
\arguments{
\item{experiment}{Name of the experiment}

\item{task}{Name of the task}

\item{id}{Id number}

\item{email}{Email to send updates to}

\item{walltime}{Requested walltime}

\item{queue}{Name of the queue to use}

\item{template}{Contents of a whisker template, or NULL to use the
default.}

\item{path}{Path to write the pbs file to.}
}
\description{
Generate a pbs file from a template.  This is extremely unlikely
to be useful unless you are me.
}

