% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgrath.R
\docType{data}
\name{mcgrath}
\alias{mcgrath}
\title{Data from "Chocolate scents and product sales: a randomized controlled trial in a Canadian bookstore and cafe"}
\format{
A data frame with 30 rows and 5 variables:
\describe{
  \item{book}{Sales of books}
  \item{coffee}{Sales of bulk coffee, tea, or spices}
  \item{food}{Sales of food}
  \item{grandtotal}{Total of book, coffee, and food sales}
  \item{treatment}{Treatment indicator. 1 = treatment, 0 = control}
}
}
\source{
\url{https://springerplus.springeropen.com/articles/10.1186/s40064-016-2303-5}
}
\usage{
mcgrath
}
\description{
This dataset is used in the exercises of Chapter 6 from the textbook.
}
\keyword{datasets}
