% Generated by roxygen2 (4.0.2): do not edit by hand
\name{move_pbs_logs}
\alias{move_pbs_logs}
\title{Move PBS output into the approprate place}
\usage{
move_pbs_logs(path = ".", jobfile = "pbs_jobs.csv", verbose = TRUE)
}
\arguments{
\item{path}{Path to the PBS files.  Assumed to be current
directory if omitted.}

\item{jobfile}{Name of the csv file that associates PBS id numbers
with experimentr.  The default here is the same as
\code{\link{launch_pbs}}.}

\item{verbose}{Print information about what is being moved?}
}
\description{
Move PBS output into the approprate place.
}

