% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mullinix.R
\docType{data}
\name{mullinix}
\alias{mullinix}
\title{Data from "The Feedback Effects of Controversial Police Use of Force"}
\format{
A data frame with 2915 rows and 2 variables:
\describe{
  \item{FL_10_DO}{Condition indicator. L = Lethal, NL = Non-lethal}
  \item{Q30}{Outcome measure}
}
}
\source{
Mullinix, Kevin J., et al. “The Feedback Effects of Controversial Police Use of Force.” Political Behavior, vol. 43, no. 2, 2020, pp. 881–898., https://doi.org/10.1007/s11109-020-09646-x.
}
\usage{
mullinix
}
\description{
This dataset is used in the exercises of Chapters 4 and 7 from the textbook.
}
\keyword{datasets}
