% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murrar.R
\docType{data}
\name{murrar}
\alias{murrar}
\title{Data from "Entertainment-education effectively reduces prejudice"}
\format{
A data frame with 193 rows and 7 variables:
\describe{
  \item{AttWhiteT0}{Attitude toward Whites pre-treatment}
  \item{AttWhiteT1}{Attitude toward Whites post-treatment}
  \item{AttArabT0}{Attitude toward Arabs pre-treatment}
  \item{AttArabT1}{Attitude toward Arabs post-treatment}
  \item{AttArabT2}{Attitude toward Arabs 4-6 weeks after treatment}
  \item{iatT1N}{IAT toward Arabs}
  \item{cond}{Treatment identifier. 1 = treatment, 0 = control}
}
}
\source{
\url{https://journals.sagepub.com/doi/abs/10.1177/1368430216682350?journalCode=gpia}
}
\usage{
murrar
}
\description{
This dataset is used in the exercises of Chapters 4 and 7 from the textbook and Chapters 4 and 5 from the R Companion.
}
\keyword{datasets}
