% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-class.R
\name{new_experiment}
\alias{new_experiment}
\title{Setup an experiment environment}
\usage{
new_experiment(project_name, project_description, experiment_name,
  experiment_author = Sys.getenv("USERNAME"), experiment_description)
}
\arguments{
\item{project_name}{A character string giving the project name used to group experiments}

\item{project_description}{A character string describing the project in more detail}

\item{experiment_name}{A character string name of the current experiment}

\item{experiment_author}{A character string specifying who ran the experiment}

\item{experiment_description}{A character string giving a short description}
}
\value{
An experiment environment
}
\description{
Use this function to create an experiment environment that holds all information
about which experiment was run by who with what results
}
