% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickerson.R
\docType{data}
\name{nickerson}
\alias{nickerson}
\title{Data from "Is Voting Contagious? Evidence from Two Field Experiments"}
\format{
A data frame with 956 rows and 2 variables:
\describe{
  \item{treatment}{Treatment indicator denoting whether participant spoke with canvassers about the upcoming election or about recycling}
  \item{voted02p}{Whether the participant voted in the elections}
}
}
\source{
\url{https://www.jstor.org/stable/27644497}
}
\usage{
nickerson
}
\description{
This dataset is used in the exercises of Chapters 4 and 6 from the R Companion.
}
\keyword{datasets}
