% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parserDf}
\alias{parserDf}
\alias{parserDatatable}
\alias{parserList}
\title{File parsers}
\usage{
parserDf(file, ...)

parserDatatable(file, ...)

parserList(file, ...)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Path to file.}

\item{...}{[any]
Further arguments passed down to actual parse/import/read function.}
}
\value{
[any]
}
\description{
File parsers
}
