% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pbs_time}
\alias{pbs_time}
\title{Generate pbs time string}
\usage{
pbs_time(days = 1, hours = 0, minutes = 0)
}
\arguments{
\item{days}{Number of days}

\item{hours}{Number of hours}

\item{minutes}{Number of minutes (less than 60)}
}
\description{
Generate a suitable time string for pbs
}
\author{
Rich FitzJohn
}

