% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poori.R
\docType{data}
\name{poori}
\alias{poori}
\title{Data from Yadav's poori experiment}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{score}{Observed outcome}
  \item{group}{Treatment condition. 1 = treatment, 0 = control}
}
}
\source{
Yadav, Monika. “Practicum Experiment: Analysis.” Columbia University, 2020.
}
\usage{
poori
}
\description{
This dataset is used in the exercises of Chapter 3 from the textbook.
}
\keyword{datasets}
