% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_progress}
\alias{print_progress}
\alias{progress}
\title{Print job information}
\usage{
print_progress(experiment, task)

progress(experiment, task)
}
\arguments{
\item{experiment}{Name of the experiment}

\item{task}{Name of the task within experiment}
}
\description{
Print information on how many jobs have started (or possibly how
many have completed)
}

