% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rondeau.R
\docType{data}
\name{rondeau}
\alias{rondeau}
\title{Data from "Matching and Challenge Gifts to Charity:Evidence from Laboratory and Natural Field Experiments"}
\format{
A data frame with 3000 rows and 2 variables:
\describe{
  \item{treattype}{Condition indicator where 0 is High Control, 1 is Challenge, 2 is Matching, and 3 is Low Control}
  \item{donation}{Donation amount}
}
}
\source{
\url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=1136440}
}
\usage{
rondeau
}
\description{
This dataset is used in the exercises of Chapter 6 from the textbook.
}
\keyword{datasets}
