% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_task}
\alias{run_task}
\title{Run a task}
\usage{
run_task(experiment, task, id = NULL, parallel = TRUE, ...,
  dry_run = FALSE, backup = FALSE, capture_output = NULL)
}
\arguments{
\item{experiment}{Name of the experiment}

\item{task}{Name of the task}

\item{id}{Id of the jobs to run.  If omitted, all ids for this
experiment will be run.}

\item{parallel}{Run in parallel using \code{parallel::mclapply}?}

\item{...}{Arguments passed through to \code{mclapply}.  In
particular, it will probably be useful to set
\code{mc.preschedule=FALSE} and specify \code{mc.cores}.}

\item{dry_run}{Logical, indicating if we should \emph{actually}
run things.}

\item{backup}{Logical, backup any existing output}

\item{capture_output}{Should we capture output?  By default this
happens if running noninteractively, or if more than one id is
used.}
}
\description{
Run a task.  This is designed to run locally, rather than being
run from a cluster.
}

