% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setup_experiment}
\alias{setup_experiment}
\title{Set up an experiment}
\usage{
setup_experiment(path, pars, packages = NULL, scripts = NULL,
  metadata = NULL, overwrite = FALSE, purge = FALSE)
}
\arguments{
\item{path}{Path within \code{output/} that things will be stored
in.  Also the name of the key in the yml.}

\item{pars}{data.frame of parameters}

\item{packages}{Optional character vector of packages to load for
every task in this experiment}

\item{scripts}{Optional characte vector of scripts to source for
every task in this experiment}

\item{metadata}{Metadata hook}

\item{overwrite}{Overwrite an existing experiment?}

\item{purge}{If overwriting, also delete all the output from that
experiment?}
}
\description{
Set up an experiment
}
\author{
Rich FitzJohn
}

