% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherman.R
\docType{data}
\name{sherman}
\alias{sherman}
\title{Data from "Deterrent Effects of Police Raids on Crack Houses: A Randomized, Controlled Experiment"}
\format{
A data frame with 207 rows and 2 variables:
\describe{
  \item{treated}{Treatment indicator. 1 = treatment, 0 = control}
  \item{CFS_after_5weeks}{Average crime reports in 5 weeks after treatment}
}
}
\source{
\url{https://www.tandfonline.com/doi/abs/10.1080/07418829500096281}
}
\usage{
sherman
}
\description{
This dataset is used in the exercises of Chapter 4 from the textbook.
}
\keyword{datasets}
