% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\docType{data}
\name{youtube}
\alias{youtube}
\title{Data from Rahmani's YouTube experiment}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{group}{Treatment condition. 1 = treatment, 0 = control}
  \item{outcome}{Observed outcome}
}
}
\usage{
youtube
}
\description{
This dataset is used in the exercises of Chapter 3 from the textbook.
}
\keyword{datasets}
