% Generated by roxygen2 (4.0.2): do not edit by hand
\name{main}
\alias{main}
\title{Run an experiment, from the command line}
\usage{
main(args = NULL, dry_run = FALSE)
}
\arguments{
\item{args}{Arguments (or NULL to read from R's \code{commandArgs})}

\item{dry_run}{Just print information but don't run anything}
}
\description{
Run an experiment, from the command line.
}
\details{
Don't use this in interactive mode this won't work
}

