% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickerson.R
\docType{data}
\name{nickerson}
\alias{nickerson}
\title{Data from "Is Voting Contagious? Evidence from Two Field Experiments"}
\format{
A data frame with 956 rows and 2 variables:
\describe{
  \item{treatment}{Treatment indicator denoting whether participant spoke with canvassers about the upcoming election or about recycling}
  \item{house_contacted}{MISSING DESCRIPTION OR VARIABLE NOT EXCLUDED}
  \item{voted02p}{Whether the participant voted in the elections}
  \item{voted00}{MISSING DESCRIPTION OR VARIABLE NOT EXCLUDED}
}
}
\source{
NICKERSON, DAVID W. “Is Voting Contagious? Evidence from Two Field Experiments.” American Political Science Review, vol. 102, no. 1, 2008, pp. 49–57., doi:10.1017/S0003055408080039.
}
\usage{
nickerson
}
\description{
This dataset is used in the exercises of Chapters 4 and 6 from the R Companion.
}
\keyword{datasets}
