% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickerson.R
\docType{data}
\name{nickerson}
\alias{nickerson}
\title{Data from "Is Voting Contagious? Evidence from Two Field Experiments"}
\format{
A data frame with 956 rows and 2 variables:
\describe{
  \item{treatment}{Treatment indicator denoting whether participant spoke with canvassers about the upcoming election or about recycling}
  \item{house_contacted}{MISSING DESCRIPTION OR VARIABLE NOT EXCLUDED}
  \item{voted02p}{Whether the participant voted in the elections}
  \item{voted00}{MISSING DESCRIPTION OR VARIABLE NOT EXCLUDED}
}
}
\source{
\url{https://www.cambridge.org/core/journals/american-political-science-review/article/abs/is-voting-contagious-evidence-from-two-field-experiments/8C2E64552D946C87FD062DD2CCD9054E}
}
\usage{
nickerson
}
\description{
This dataset is used in the exercises of Chapters 4 and 6 from the R Companion.
}
\keyword{datasets}
