% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrenlewis.R
\docType{data}
\name{wrenlewis}
\alias{wrenlewis}
\title{Data from "Formalizing land rights can reduce forest loss: Experimental evidence from Benin"}
\format{
A data frame with 102 rows and 3 variables:
\describe{
  \item{treated}{Treatment indicator. 1 = treatment, 0 = control}
  \item{tcl_area_L_from2010to2017}{Hectares of forest loss 2010-2017}
  \item{tcl_area_L_from2001to2009}{Hectares of forest loss 2001-2009}
}
}
\source{
Wren-Lewis, Liam et al. “Formalizing land rights can reduce forest loss: Experimental evidence from Benin.” Science advances vol. 6,26 eabb6914. 26 Jun. 2020, doi:10.1126/sciadv.abb6914
}
\usage{
wrenlewis
}
\description{
This dataset is used in the exercises of Chapter 4 from the textbook.
}
\keyword{datasets}
